Added code validation and valid bit to letter import
This marks letters with an invalid letter code and thus can exclude it
from the date range request for example.
Note that this is not as strict as the pattern demanded by the GUI as
there are many values in the dbf file that have no decimal places.
It also tries to correct decimal places in the form of .101 to .11 to
accommodate the new code structure.