Implemented filter groups
The heavy lifting is done. Search requests are parsed and converted
into a structure of at most binary filters, that can be easily
translated into search querys.
The GUI is modified to allow the creation of filter groups and choose
the filter junction (or, and).
There is a lot of work to do to create a good user experience, but the
first step is done.

Note: This structure would also allow the addition of something like a
LetterFilter, that could realize the match according to properties on
the letter row.